requirements.txt
pymongo>=4.0.0
python-dotenv>=0.19.0
pymongo[encryption]